requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
asyncpg==0.29.0
attrs==23.2.0
bidict==0.23.1
black==24.4.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
docstring_parser==0.16
docutils==0.19
email_validator==2.1.1
executing==2.0.1
fastapi==0.109.2
fastapi-cli==0.0.3
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
Hypercorn==0.16.0
hyperframe==6.0.1
icecream==2.1.3
idna==3.7
ifaddr==0.2.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nicegui==1.4.25
orjson==3.10.3
packaging==24.0
pathspec==0.12.1
piccolo==1.5.0
piccolo_admin==1.3.3
piccolo_api==1.4.0
platformdirs==4.2.2
priority==2.0.0
pscript==0.7.7
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
targ==0.4.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
uvicorn==0.29.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
